requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
PyPDF2
faiss-cpu
langchain_google_genai
PyMuPDF 